storage: move migration to own file
